+++ src/main/java/racingcar/domain/Engine.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import racingcar.random.NumberGenerator;
